MAINT: Moved a netbox api method into the netbox_data module.
